+++ src/main.tsx
@@ -1,5 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
+import { HashRouter } from 'react-router-dom'
 import { ThemeProvider } from '@/contexts/theme-context'
 import { SidebarProvider } from '@/contexts/sidebar-context'
 import './index.css'
@@ -11,12 +12,15 @@ if (rootElement === null) {
   throw new Error('Root element not found')
 }
 
+/* HashRouter para compatibilidade com GitHub Pages (SSG) */
 createRoot(rootElement).render(
   <StrictMode>
-    <ThemeProvider>
-      <SidebarProvider>
-        <App />
-      </SidebarProvider>
-    </ThemeProvider>
+    <HashRouter>
+      <ThemeProvider>
+        <SidebarProvider>
+          <App />
+        </SidebarProvider>
+      </ThemeProvider>
+    </HashRouter>
   </StrictMode>
 )